Make a backup of our original CR
This section isn't exactly idempotent, so we need a way to get our
original file back. For now, we just copy the original to its own dir so
the rest of the playbook can run. Ideally, we return the repo back to
its original state entirely.

TODO: Make this closer to idempotent