Internal registries expanded (#25)
* Lay out the vars we need to populate

* Get the deployments, image names, and image versions

Same way we did it for redis.

TODO: We should be able to do this in a loop for DRY

* Remove components which use the same image

* Remove extra line break

* Add our new images to the template

* Make a backup of our original CR

This section isn't exactly idempotent, so we need a way to get our
original file back. For now, we just copy the original to its own dir so
the rest of the playbook can run. Ideally, we return the repo back to
its original state entirely.

TODO: Make this closer to idempotent

* Add missing braces

* copy module doesn't use 'state'